requirements.txt
numpy
pillow
scikit-learn
imbalanced-learn
streamlit
joblib
